move CarliniWagnerL2 to a separate file
